+++ src/Maestro/Maestro.Web.Tests/BuildController20200914Tests.cs
@@ -7,6 +7,7 @@ using Maestro.Web.Api.v2020_02_20.Controllers;
 using Maestro.Web.Api.v2020_02_20.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.DotNet.DarcLib;
+using Microsoft.DotNet.Internal.Testing.DependencyInjection.Abstractions;
 using Microsoft.DotNet.Internal.Testing.Utility;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
@@ -21,7 +22,7 @@ using Commit = Maestro.Web.Api.v2020_02_20.Models.Commit;
 namespace Maestro.Web.Tests
 {
     [TestFixture]
-    public class BuildController20200914Tests
+    public partial class BuildController20200914Tests
     {
         static string repository = "FAKE-REPOSITORY";
         static string commitHash = "FAKE-COMMIT";
@@ -34,7 +35,7 @@ namespace Maestro.Web.Tests
         [Test]
         public async Task CommitIsFound()
         {
-            using TestData data = await BuildDefaultAsync();
+            using TestData data = await TestData.Default.BuildAsync();
 
             int id;
             {
@@ -77,18 +78,12 @@ namespace Maestro.Web.Tests
             }
         }
 
-        private Task<TestData> BuildDefaultAsync()
+        [TestDependencyInjectionSetup]
+        private static class TestDataConfiguration
         {
-            return new TestDataBuilder().BuildAsync();
-        }
-
-        private sealed class TestDataBuilder
-        {
-            public async Task<TestData> BuildAsync()
+            public static async Task Dependencies(IServiceCollection collection)
             {
                 string connectionString = await SharedData.Database.GetConnectionString();
-
-                ServiceCollection collection = new ServiceCollection();
                 collection.AddLogging(l => l.AddProvider(new NUnitLogger()));
 
                 var mockIRemoteFactory = new Mock<IRemoteFactory>();
@@ -108,29 +103,13 @@ namespace Maestro.Web.Tests
                     options.UseSqlServer(connectionString);
                     options.EnableServiceProviderCaching(false);
                 });
-                collection.AddTransient<BuildsController>();
                 collection.AddSingleton<ISystemClock, TestClock>();
-                ServiceProvider provider = collection.BuildServiceProvider();
-
-                var controller = provider.GetRequiredService<BuildsController>();
-
-                return new TestData(provider, controller);
-            }
-        }
-        private sealed class TestData : IDisposable
-        {
-            private readonly ServiceProvider _provider;
-            public BuildsController Controller { get; }
-
-            public TestData(ServiceProvider provider, BuildsController controller)
-            {
-                _provider = provider;
-                Controller = controller;
             }
 
-            public void Dispose()
+            public static Func<IServiceProvider, BuildsController> Controller(IServiceCollection collection)
             {
-                _provider.Dispose();
+                collection.AddTransient<BuildsController>();
+                return s => s.GetRequiredService<BuildsController>();
             }
         }
     }
