+++ src/Maestro/Maestro.Web.Tests/ChannelsController20200220Tests.cs
@@ -9,6 +9,7 @@ using Maestro.Web.Api.v2020_02_20.Controllers;
 using Maestro.Web.Api.v2020_02_20.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.DotNet.DarcLib;
+using Microsoft.DotNet.Internal.Testing.DependencyInjection.Abstractions;
 using Microsoft.DotNet.Internal.Testing.Utility;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
@@ -22,12 +23,12 @@ using NUnit.Framework;
 namespace Maestro.Web.Tests
 {
     [TestFixture]
-    public class ChannelsController20200220Tests
+    public partial class ChannelsController20200220Tests
     {
         [Test]
         public async Task CreateChannel()
         {
-            using TestData data = await BuildDefaultAsync();
+            using TestData data = await TestData.Default.BuildAsync();
             Channel channel;
             string channelName = "TEST-CHANNEL-BASIC";
             string classification = "TEST-CLASSIFICATION";
@@ -57,7 +58,7 @@ namespace Maestro.Web.Tests
         [Test]
         public async Task ListRepositories()
         {
-            using TestData data = await BuildDefaultAsync();
+            using TestData data = await TestData.Default.BuildAsync();
             string channelName = "TEST-CHANNEL-LIST-REPOSITORIES";
             string classification = "TEST-CLASSIFICATION";
             string commitHash = "FAKE-COMMIT";
@@ -105,7 +106,7 @@ namespace Maestro.Web.Tests
         [Test]
         public async Task AddingBuildToChannelTwiceWorks()
         {
-            using TestData data = await BuildDefaultAsync();
+            using TestData data = await TestData.Default.BuildAsync();
             const string channelName = "TEST-CHANNEL-ADD-TWICE-2020";
             const string classification = "TEST-CLASSIFICATION";
             const string commitHash = "FAKE-COMMIT";
@@ -150,26 +151,12 @@ namespace Maestro.Web.Tests
             }
         }
 
-        private Task<TestData> BuildDefaultAsync()
+        [TestDependencyInjectionSetup]
+        private static class TestDataConfiguration
         {
-            return new TestDataBuilder().BuildAsync();
-        }
-
-        private sealed class TestDataBuilder
-        {
-            private Type _backgroundQueueType = typeof(NeverBackgroundQueue);
-
-            public TestDataBuilder WithImmediateBackgroundQueue()
-            {
-                _backgroundQueueType = typeof(ImmediateBackgroundQueue);
-                return this;
-            }
-
-            public async Task<TestData> BuildAsync()
+            public static async Task Dependencies(IServiceCollection collection)
             {
                 string connectionString = await SharedData.Database.GetConnectionString();
-
-                var collection = new ServiceCollection();
                 collection.AddLogging(l => l.AddProvider(new NUnitLogger()));
                 collection.AddSingleton<IHostEnvironment>(new HostingEnvironment
                 {
@@ -180,36 +167,26 @@ namespace Maestro.Web.Tests
                     options.UseSqlServer(connectionString);
                     options.EnableServiceProviderCaching(false);
                 });
-                collection.AddSingleton<ChannelsController>();
-                collection.AddSingleton<BuildsController>();
-                collection.AddSingleton<ISystemClock, TestClock>();
                 collection.AddSingleton(Mock.Of<IRemoteFactory>());
-                collection.AddSingleton(typeof(IBackgroundQueue), _backgroundQueueType);
-                ServiceProvider provider = collection.BuildServiceProvider();
-
-                var clock = (TestClock) provider.GetRequiredService<ISystemClock>();
-
-                return new TestData(provider, clock);
+                collection.AddSingleton<IBackgroundQueue, NeverBackgroundQueue>();
             }
-        }
 
-        private sealed class TestData : IDisposable
-        {
-            private readonly ServiceProvider _provider;
-            public TestClock Clock { get; }
-
-            public TestData(ServiceProvider provider, TestClock clock)
+            public static Func<IServiceProvider, TestClock> Clock(IServiceCollection collection)
             {
-                _provider = provider;
-                Clock = clock;
+                collection.AddSingleton<ISystemClock, TestClock>();
+                return s=> (TestClock) s.GetRequiredService<ISystemClock>();
             }
-            
-            public ChannelsController Controller => _provider.GetRequiredService<ChannelsController>();
-            public BuildsController BuildsController => _provider.GetRequiredService<BuildsController>();
 
-            public void Dispose()
+            public static Func<IServiceProvider, ChannelsController> Controller(IServiceCollection collection)
             {
-                _provider.Dispose();
+                collection.AddSingleton<ChannelsController>();
+                return s=> s.GetRequiredService<ChannelsController>();
+            }
+
+            public static Func<IServiceProvider, BuildsController> BuildsController(IServiceCollection collection)
+            {
+                collection.AddSingleton<BuildsController>();
+                return s=> s.GetRequiredService<BuildsController>();
             }
         }
     }
