+++ src/Maestro/Maestro.Web/Maestro.Web.csproj
@@ -3,7 +3,6 @@
   <PropertyGroup>
     <IsServiceFabricServiceProject>True</IsServiceFabricServiceProject>
     <NoWarn>$(NoWarn);1591</NoWarn>
-    <RuntimeIdentifier>win-x64</RuntimeIdentifier>
     <SignAssembly>false</SignAssembly>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
   </PropertyGroup>
