+++ src/Maestro/Microsoft.AspNetCore.ApiVersioning/Microsoft.AspNetCore.ApiVersioning.csproj
@@ -1,6 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
+    <RuntimeIdentifier></RuntimeIdentifier>
     <IsPackable>true</IsPackable>
   </PropertyGroup>
