+++ src/Microsoft.DotNet.ServiceFabric.ServiceHost/Microsoft.DotNet.ServiceFabric.ServiceHost.csproj
@@ -1,6 +1,5 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <RuntimeIdentifier>win-x64</RuntimeIdentifier>
     <IsShipping>false</IsShipping>
     <IsPackable>true</IsPackable>
   </PropertyGroup>
