+++ src/Shared/Microsoft.DotNet.Internal.Testing.DependencyInjection.Abstractions/Microsoft.DotNet.Internal.Testing.DependencyInjection.Abstractions.csproj
@@ -1,3 +1,2 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
 </Project>
