+++ src/Shared/Microsoft.DotNet.Internal.Testing.DependencyInjectionCodeGen/Microsoft.DotNet.Internal.Testing.DependencyInjectionCodeGen.csproj
@@ -1,6 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
+    <RuntimeIdentifier></RuntimeIdentifier>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.CodeAnalysis.Analyzers" PrivateAssets="all" />
